move PDF specific code out to own file
