+++ Tools/CmdLine/IccProfileVisualize/MiniPDF.hpp
@@ -67,88 +67,9 @@
 #include <vector>
 #include <string>
 #include <map>
+#include "vizShared.hpp"
 
 /******************************************************************************/
-
-const float inch2mm = 25.4f;                 // 25.4 millimeters per inch, international standard
-const float mm2point = 72.0f / inch2mm;      // 2.834645669 (Shows up severeal places)
-const float inch2point = 72.0f;              // 72 points per inch, DTP and W3C standard
-
-
-struct point2D {
-  point2D(float xx, float yy) : x(xx), y(yy) {}
-  point2D() : x(0.0), y(0.0) {}
-  
-  bool operator<(const point2D& o) const {
-  if (x == o.x)
-    return y < o.y;
-  else
-    return x < o.x;
-  }
-
-  float x, y;
-};
-
-inline point2D operator+(const point2D& xx, const point2D& yy) {
-  return point2D(xx.x + yy.x, xx.y + yy.y);
-}
-
-inline point2D operator-(const point2D& xx, const point2D& yy) {
-  return point2D(xx.x - yy.x, xx.y - yy.y);
-}
-
-inline point2D& operator+=(point2D& xx, const point2D& yy) {
-  xx.x += yy.x;
-  xx.y += yy.y;
-  return xx;
-}
-
-inline point2D& operator-=(point2D& xx, const point2D& yy) {
-  xx.x -= yy.x;
-  xx.y -= yy.y;
-  return xx;
-}
-
-inline point2D operator*(const point2D& xx, const point2D& yy) {
-  return point2D(xx.x * yy.x, xx.y * yy.y);
-}
-
-inline point2D operator*(const point2D& xx, const float ss) {
-  return point2D(xx.x * ss, xx.y * ss);
-}
-
-inline point2D operator/(const point2D& xx, const float ss) {
-  return point2D(xx.x / ss, xx.y / ss);
-}
-
-inline point2D operator*(const float ss, const point2D& yy) {
-  return point2D(ss * yy.x, ss * yy.y);
-}
-
-inline point2D operator/(const point2D& xx, const point2D& yy) {
-  return point2D(xx.x / yy.x, xx.y / yy.y);
-}
-
-typedef std::vector<point2D> pointList;
-
-/******************************************************************************/
-
-struct Rect2D {
-  Rect2D(float leftin, float rightin, float bottomin, float topin) :
-    left(leftin), right(rightin), bottom(bottomin), top(topin)
-    {}
-
-  Rect2D() : left(0.0f), right(0.0f), bottom(0.0f), top(0.0f)
-    {}
-
-  Rect2D(const point2D &ll, const point2D &tr) :
-    left(ll.x), right(tr.x), bottom(ll.y), top(tr.y)
-    {}
-
-  point2D Size() const  { return point2D( right-left, bottom-top ); }
-
-  float left, right, bottom, top;
-};
 
 std::ostream& operator<<( std::ostream &os, const Rect2D &r );
 std::ostream& operator<<( std::ostream &os, const point2D &r );
